New approach
The tree is not recursively defined anymore. There's a TreeNode class
that keeps the data and the pointers to the other nodes in the tree, the
Tree object only keeps the root of the tree.

All methods have been shifted to the TreeNode class, a proper
destructor has been written for the Tree class.