New approach
The HashPair is now a struct and not a class anymore (it wasn't
necessary for the working of the data structure).

HashList class has been removed. Instead the list holds an abstract type
ValueType that is a key-pair value by default, and is the type of the
key if the type of the value is void. (This has been used to implement a
Set based on a HashTable).

The type of the value held in the hash table is now an abstract
HashTable type. This way the insert and contains method are the same
even if the second template argument is void.

The iterator class has been rewritten to be part of the class as a
nested class and makes use of the iterator class of the list type.