New Approach
The three types of set are not in their separate namespace anymore.
Instead each is a class of its own.

The set based on unordered list is now called UnorderedListSet and has
been modified to work with the new List data structure approach. The
iterator for the class is an alias for the List::iterator class.

The set_iterator classes for the other two types have been removed as
well, as they're now aliases for the iterator of the underlying data
structures.

The set data structure is now STL compliant.